Delete the score increments outside of playRound
The score increments that were outside of the
playRound function and inside the event listeners
for each button choice was causing a bug in the
code that would double the score of a player at
the end of each round. Getting rid of the extra
incrments ensures that each score is incremented
once.